### 2017-06-20 #297 @ashleefv
The science has already been peer-reviewed and published this spring:
http://www.sciencedirect.com/science/article/pii/S009813541730145X. The
code also includes stand alone .m files that can be viewed in any editor.

Ashlee N. Ford Versypt, Ph.D.
Assistant Professor, School of Chemical Engineering, Oklahoma State
University
Member, Harold Hamm Diabetes Center, University of Oklahoma Health Sciences
Center
420 Engineering North, OSU, Stillwater, OK 74078
Ph: 405-744-5280
ashleefv@okstate.edu
Ford Versypt Lab Website <https://sites.google.com/site/ashleefordversypt/>

On Tue, Jun 20, 2017 at 12:24 PM, Lorena A. Barba <notifications@github.com>
wrote:

> It's quite removed from any expertise of mine. I have no idea how to even
> begin finding a reviewer for this. (Not to mention it's a GUI app for
> MATLAB.)
>
