### 2017-06-20 #297 @Kevin-Mattheus-Moerman
Not my field either but myself and @cMadan are MATLAB experts so perhaps we can have a look here too.
