### 2017-06-20 #297 @Kevin-Mattheus-Moerman
The [readme.md](https://github.com/ashleefv/ACEInhibPKPD/blob/master/README.md) and [paper.md](https://github.com/ashleefv/ACEInhibPKPD/blob/master/paper.md) emphasize "This software is intended to be used by engineering educators working with college freshmen or high school students and by systems biology researchers investigating pharmaceuticals to modulate the renin-angiotensin system.". Can you elaborate on the scientific research applications a bit more in the paper?
Also, there is a small typo in the paper "...describe and solve the the dynamics of...".
