### 2017-06-21 #297 @ashleefv
I updated both the readme.md and paper.md with additional statements about
the scientific applications. Is further clarification needed?

Ashlee N. Ford Versypt, Ph.D.
Assistant Professor, School of Chemical Engineering, Oklahoma State
University
Member, Harold Hamm Diabetes Center, University of Oklahoma Health Sciences
Center
420 Engineering North, OSU, Stillwater, OK 74078
Ph: 405-744-5280
ashleefv@okstate.edu
Ford Versypt Lab Website <https://sites.google.com/site/ashleefordversypt/>

On Tue, Jun 20, 2017 at 3:51 PM, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> The readme.md
> <https://github.com/ashleefv/ACEInhibPKPD/blob/master/README.md> and
> paper.md <https://github.com/ashleefv/ACEInhibPKPD/blob/master/paper.md>
> emphasize "This software is intended to be used by engineering educators
> working with college freshmen or high school students and by systems
> biology researchers investigating pharmaceuticals to modulate the
> renin-angiotensin system.". Can you elaborate on the scientific research
> applications a bit more in the paper?
> Also, there is a small typo in the paper "...describe and solve the the
> dynamics of...".
>
