### 2017-06-21 #297 @Kevin-Mattheus-Moerman
Thanks for making the changes. I'll review these and let you know. Although I am happy implementing/reviewing MATLAB code this is not my field of study. Can you recommend any reviewers (e.g. researchers that would be interested in your work but were not involved in its creation)?
