### 2017-06-21 #297 @Kevin-Mattheus-Moerman
@arfon @labarba I can take this as editor
