### 2017-06-21 #297 @ashleefv
potential reviewers:

   - Fulden Buyukozturk
   https://www.linkedin.com/in/fulden-buyukozturk-b4384910/
   - Gilbert Koch https://www.ukbb.ch/en/personal/personen/koch-gilbert.php

Ashlee N. Ford Versypt, Ph.D.
Assistant Professor, School of Chemical Engineering, Oklahoma State
University
Member, Harold Hamm Diabetes Center, University of Oklahoma Health Sciences
Center
420 Engineering North, OSU, Stillwater, OK 74078
Ph: 405-744-5280
ashleefv@okstate.edu
Ford Versypt Lab Website <https://sites.google.com/site/ashleefordversypt/>

On Wed, Jun 21, 2017 at 3:16 PM, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> Thanks for making the changes. I'll review these and let you know.
> Although I am happy implementing/reviewing MATLAB code this is not my field
> of study. Can you recommend any reviewers (e.g. researchers that would be
> interested in your work but were not involved in its creation)?
>
