### 2017-06-22 #297 @dpastoor
@Kevin-Mattheus-Moerman Thanks for your interest. First, I would like to commend @ashleefv for putting her research in the open. This is a fantastic first step, however the core functionality is written in matlab, which is not FOSS. Looking at the guidelines, this does not seem to directly violate the journal requirements, however as I do not have a copy of matlab I will not be able to review.

I will encourage @ashleefv to consider using a fully open source implementation, such as using R shiny for the GUI + https://github.com/metrumresearchgroup/mrgsolve for the PKPD solver or something along those lines, at which point I would be happy to review
