### 2017-06-22 #297 @Kevin-Mattheus-Moerman
@dpastoor thanks for your fast reply. I understand your points. We do not exclude works in principle that require proprietor hardware or software (e.g. a runtime like MATLAB). As long as it adheres to the OSI definitions of open source.
If we were able to set-up a free to use MATLAB environment (have to enquire), would you consider reviewing this work?
