### 2017-06-22 #297 @Kevin-Mattheus-Moerman
@ashleefv can you briefly inquire if translation to something like [octave](https://www.gnu.org/software/octave/) is easy to achieve? Translation is NOT a requirement but it would be good to state here if you believe it to be achievable in principle.
