### 2017-06-22 #297 @dpastoor
@Kevin-Mattheus-Moerman I have 2 problems reviewing this (I probably _could_ get ahold of matlab myself).

First and foremost, as I a have not written any matlab, I would not feel comfortable evaluating the implementation, which I believe is an important part of any scientific review. I can handle R, python, go, or javascript implementations, if at some point it is translated to any of those languages.

Second, I appreciate that matlab is extremely prevalent in engineering, and certain components, such as a GUI, were previously much more labor intensive to implement in FOSS languages. However, in this day and age, it is not an unreasonable task. To me, the thing that excites me seeing this journal is the ability to provide a "reviewed" piece of software that anyone, anywhere, can immediately pull and start using. To that end, I would be willing to 'donate' my time to support such software efforts, however in tandem with the first point above I cannot justify the bandwidth needed for a "closed" matlab implementation to take the time to properly review.  Again, not to take anything from @ashleefv 's willingness to at least put the code in the open, which I think is *amazing*.

If @ashleefv is willing to do a translation, I would be happy to assist in both the translation and/or review.
