### 2017-06-22 #297 @ashleefv
I looked into octave. Unfortunately, even for the code that doesnt need the
GUI, octave cannot handle the .mat files used for storing data and
variables for various cases.

On Jun 22, 2017 9:57 AM, "Kevin Mattheus Moerman" <notifications@github.com>
wrote:

> @ashleefv <https://github.com/ashleefv> can you briefly inquire if
> translation to something like octave
> <https://www.gnu.org/software/octave/> is easy to achieve? Translation is
> NOT a requirement but it would be good to state here if you believe it to
> be achievable in principle.
>
