### 2017-06-22 #297 @Kevin-Mattheus-Moerman
@dpastoor Thanks for your detailed reply. I understand your points. We would be happy to have you as reviewer on other projects. And here is a list of currently open reviews:
https://github.com/openjournals/joss-reviews/issues/
