### 2017-07-17 #297 @Kevin-Mattheus-Moerman
:wave: @tpfau @sdelandtsheer would you willing to review this JOSS submission?

Here is the [JOSS](http://joss.theoj.org/) website
Here is the repository for this project: https://github.com/ashleefv/ACEInhibPKPD
Reviewing JOSS submissions generally does not take a lot of time and is largely focused on the claims and functionality of the software. See here for [reviewer instructions](http://joss.theoj.org/about#reviewer_guidelines)

Let me know if you have any questions.
