### 2017-07-30 #297 @Kevin-Mattheus-Moerman
@tpfau @sdelandtsheer would you willing to review this JOSS submission?
