### 2017-07-30 #297 @ashleefv
They may not be on Github. If needed because of no reviewers, I can
withdraw this submission.

Ashlee N. Ford Versypt, Ph.D.
Assistant Professor, School of Chemical Engineering, Oklahoma State
University
Member, Harold Hamm Diabetes Center, University of Oklahoma Health Sciences
Center
420 Engineering North, OSU, Stillwater, OK 74078
Ph: 405-744-5280
ashleefv@okstate.edu
Ford Versypt Lab Website <https://sites.google.com/site/ashleefordversypt/>

On Sun, Jul 30, 2017 at 10:57 AM, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @ashleefv <https://github.com/ashleefv> you suggested reviewers but I
> have not find them on GitHub. Can you suggest reviewers by posting GitHub
> handles here?
>
