### 2017-07-30 #297 @ashleefv
gilbert.koch@ukbb.ch

@fulden_b on twitter

Ashlee N. Ford Versypt, Ph.D.
Assistant Professor, School of Chemical Engineering, Oklahoma State
University
Member, Harold Hamm Diabetes Center, University of Oklahoma Health Sciences
Center
420 Engineering North, OSU, Stillwater, OK 74078
Ph: 405-744-5280
ashleefv@okstate.edu
Ford Versypt Lab Website <https://sites.google.com/site/ashleefordversypt/>

On Sun, Jul 30, 2017 at 3:24 PM, Daniel S. Katz <notifications@github.com>
wrote:

> Or you can provide email addresses, and they can be asked to join GitHub
> to do the review.
>
