### 2017-07-31 #297 @Kevin-Mattheus-Moerman
Ok @ashleefv @sdelandtsheer review will happen here #340
