### 2017-07-31 #340 @sdelandtsheer
Hi @Kevin-Mattheus-Moerman I have one remark:
There is documentation missing for the function _run_PKPD_without_GUI_. The input/output should be specified clearly, this is the function to run the simulations from the command line. I also recommend making the last three arguments optional.
The app is otherwise well documented, functional, and can be accepted when this is corrected as far as I am concerned.
