### 2017-08-01 #340 @Kevin-Mattheus-Moerman
Thanks @sdelandtsheer for the fast review. @ashleefv can you work on these comments and formulate a reply?
