### 2017-08-12 #340 @arfon
> @Kevin-Mattheus-Moerman @danielskatz - it looks like the review checklist is missing from this thread. Did one of you accidentally delete it?

Well that's very strange. The top of the issue thread wasn't rendering for me just before. Ignore me!
