### 2017-08-12 d68f179
edited documentation to specify input/output clearly and to make the last 3 arguments optional
