### 2017-08-12 #340 @ashleefv
I made the requested changes in *run_PKPD_without_GUI*.m and committed
those changes.

Ashlee N. Ford Versypt, Ph.D.
Assistant Professor, School of Chemical Engineering, Oklahoma State
University
Member, Harold Hamm Diabetes Center, University of Oklahoma Health Sciences
Center
420 Engineering North, OSU, Stillwater, OK 74078
Ph: 405-744-5280
ashleefv@okstate.edu
Ford Versypt Lab Website <https://sites.google.com/site/ashleefordversypt/>

On Sat, Aug 12, 2017 at 2:22 PM, Arfon Smith <notifications@github.com>
wrote:

> @Kevin-Mattheus-Moerman <https://github.com/kevin-mattheus-moerman>
> @danielskatz <https://github.com/danielskatz> - it looks like the review
> checklist is missing from this thread. Did one of you accidentally delete
> it?
>
> Well that's very strange. The top of the issue thread wasn't rendering for
> me just before. Ignore me!
>
