### 2017-08-12 #340 @ashleefv
@Kevin-Mattheus-Moerman <https://github.com/kevin-mattheus-moerman>
@sdelandtsheer <https://github.com/sdelandtsheer> I made the requested
changes in *run_PKPD_without_GUI*.m and committed those changes.

Ashlee N. Ford Versypt, Ph.D.
Assistant Professor, School of Chemical Engineering, Oklahoma State
University
Member, Harold Hamm Diabetes Center, University of Oklahoma Health Sciences
Center
420 Engineering North, OSU, Stillwater, OK 74078
Ph: 405-744-5280
ashleefv@okstate.edu
Ford Versypt Lab Website <https://sites.google.com/site/ashleefordversypt/>

On Sat, Aug 12, 2017 at 4:06 PM, Ford Versypt, Ashlee <ashleefv@okstate.edu>
wrote:

> I made the requested changes in *run_PKPD_without_GUI*.m and committed
> those changes.
>
> Ashlee N. Ford Versypt, Ph.D.
> Assistant Professor, School of Chemical Engineering, Oklahoma State
> University
> Member, Harold Hamm Diabetes Center, University of Oklahoma Health
> Sciences Center
> 420 Engineering North, OSU, Stillwater, OK 74078
> Ph: 405-744-5280 <(405)%20744-5280>
> ashleefv@okstate.edu
> Ford Versypt Lab Website
> <https://sites.google.com/site/ashleefordversypt/>
>
>
>
> On Sat, Aug 12, 2017 at 2:22 PM, Arfon Smith <notifications@github.com>
> wrote:
>
>> @Kevin-Mattheus-Moerman <https://github.com/kevin-mattheus-moerman>
>> @danielskatz <https://github.com/danielskatz> - it looks like the review
>> checklist is missing from this thread. Did one of you accidentally delete
>> it?
>>
>> Well that's very strange. The top of the issue thread wasn't rendering
>> for me just before. Ignore me!
>>
>> —
>> You are receiving this because you were mentioned.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/340#issuecomment-322000742>,
>> or mute the thread
>> <https://github.com/notifications/unsubscribe-auth/AMD-NqSwiGcFxD1xDsST4HTXe3LHXIWAks5sXft4gaJpZM4OoTbS>
>> .
>>
>
>
