### 2017-08-13 #340 @Kevin-Mattheus-Moerman
Thanks @ashleefv. @sdelandtsheer are you happy with how your comments were addressed?
