### 2017-08-18 #340 @sdelandtsheer
Thanks @ashleefv for making the changes. This is now a well-documented function. @Kevin-Mattheus-Moerman I am happy indeed.
