### 2017-08-22 #340 @sdelandtsheer
@Kevin-Mattheus-Moerman Happy to help. Thank you for considering Matlab code in your journal. Feel free to contact me if needed.
