### 2017-09-05 #340 @arfon
I note there is already a DOI present in the README but this doesn't seem to have the JOSS-specific stuff present (e.g. the `paper.md` and `paper.bib` files).
