### 2017-09-05 c3cf4d7
update README with v1.1 DOI
