### 2017-09-05 #340 @ashleefv
@arfon I updated the DOI with the revisions and the paper.md and paper.bib
files.

https://doi.org/10.5281/zenodo.885362

Ashlee N. Ford Versypt, Ph.D.
Assistant Professor, School of Chemical Engineering, Oklahoma State
University
Member, Harold Hamm Diabetes Center, University of Oklahoma Health Sciences
Center
420 Engineering North, OSU, Stillwater, OK 74078
Ph: 405-744-5280
ashleefv@okstate.edu
Ford Versypt Lab Website <https://sites.google.com/site/ashleefordversypt/>

On Tue, Sep 5, 2017 at 9:03 AM, Arfon Smith <notifications@github.com>
wrote:

> @ashleefv <https://github.com/ashleefv> - At this point could you make an
> archive of the reviewed software in Zenodo/figshare/other service and
> update this thread with the DOI of the archive? I can then move forward
> with accepting the submission.
>
> I note there is already a DOI present in the README but this doesn't seem
> to have the JOSS-specific stuff present (e.g. the paper.md and paper.bib
> files).
>
